Split page api code
